feat: app - domain 모듈 분리
