+++ app/client/src/main/kotlin/playground/cryptobackedloans/client/DummyAppApplication.kt
@@ -1,0 +1,11 @@
+package playground.cryptobackedloans.client
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class DummyAppApplication
+
+fun main(args: Array<String>) {
+    runApplication<DummyAppApplication>(*args)
+}
