+++ app/client/src/test/kotlin/playground/cryptobackedloans/app/client/AppApplicationTests.kt
@@ -1,0 +1,13 @@
+package playground.cryptobackedloans.app.client
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class AppApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+
+}
