+++ domain/src/main/kotlin/playground/cryptobackedloans/app/client/domain/cryptoCurrency/CryptoCurrency.kt
@@ -1,0 +1,11 @@
+package playground.cryptobackedloans.app.client.domain.cryptoCurrency
+
+/**
+ * 가상자산 (BTC, ETH..)
+ */
+class CryptoCurrency(
+    val symbol: String,
+    val precision: Int,
+    val feePrecision: Int
+) {
+}
