+++ domain/src/main/kotlin/playground/cryptobackedloans/app/client/domain/loan/LoanStatus.kt
@@ -1,0 +1,8 @@
+package playground.cryptobackedloans.app.client.domain.loan
+
+enum class LoanStatus {
+    ACTIVE,
+    MARGIN_CALL,
+    LIQUIDATION,
+    REPAYMENT
+}
