+++ domain/src/main/kotlin/playground/cryptobackedloans/app/client/domain/loan/MarginCall.kt
@@ -1,0 +1,16 @@
+package playground.cryptobackedloans.app.client.domain.loan
+
+import playground.cryptobackedloans.domain.vo.Ltv
+import java.time.LocalDateTime
+
+/**
+ * 마진콜
+ */
+class MarginCall(
+    val id: Long,
+    val ltv: Ltv, // 마진콜 발동 시점의 ltv
+    val createdAt: LocalDateTime,
+    val releasedAt: LocalDateTime,
+    val updatedAt: LocalDateTime
+) {
+}
