+++ domain/src/main/kotlin/playground/cryptobackedloans/app/client/domain/transaction/Transaction.kt
@@ -1,0 +1,22 @@
+package playground.cryptobackedloans.app.client.domain.transaction
+
+import playground.cryptobackedloans.domain.vo.Quantity
+import java.time.LocalDateTime
+
+/**
+ * 입/출금 거래
+ */
+class Transaction(
+    val id: Long,
+    val cryptoCurrencyId: Long,
+    val userId: Long,
+    val collateralWalletId: Long,
+    val source: String,
+    val quantity: Quantity,
+    val networkFee: Quantity,
+    val txId: String,
+    val outputIndex: Int,
+    val createdAt: LocalDateTime,
+    val updatedAt: LocalDateTime
+) {
+}
