+++ domain/src/main/kotlin/playground/cryptobackedloans/app/client/domain/vo/Ltv.kt
@@ -1,0 +1,5 @@
+package playground.cryptobackedloans.app.client.domain.vo
+
+data class Ltv(
+    val ratio: Int
+)
