+++ domain/src/main/kotlin/playground/cryptobackedloans/app/client/domain/vo/Rate.kt
@@ -1,0 +1,5 @@
+package playground.cryptobackedloans.app.client.domain.vo
+
+data class Rate(
+    val value: Int
+)
