--- src/main/kotlin/playground/cryptobackedloans/CryptoBackedLoansApplication.kt
@@ -1,11 +1,0 @@
-package playground.cryptobackedloans
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class CryptoBackedLoansApplication
-
-fun main(args: Array<String>) {
-    runApplication<CryptoBackedLoansApplication>(*args)
-}
