--- src/main/kotlin/playground/cryptobackedloans/domain/common/vo/Ltv.kt
@@ -1,5 +1,0 @@
-package playground.cryptobackedloans.domain.common.vo
-
-data class Ltv(
-    val ratio: Int
-)
