--- src/main/kotlin/playground/cryptobackedloans/domain/common/vo/Quantity.kt
@@ -1,8 +1,0 @@
-package playground.cryptobackedloans.domain.common.vo
-
-import java.math.BigDecimal
-
-data class Quantity(
-    val value: BigDecimal
-) {
-}
