--- src/main/kotlin/playground/cryptobackedloans/domain/common/vo/Rate.kt
@@ -1,5 +1,0 @@
-package playground.cryptobackedloans.domain.common.vo
-
-data class Rate(
-    val value: Int
-)
