--- src/main/kotlin/playground/cryptobackedloans/domain/cryptoCurrency/CryptoCurrency.kt
@@ -1,11 +1,0 @@
-package playground.cryptobackedloans.domain.cryptoCurrency
-
-/**
- * 가상자산 (BTC, ETH..)
- */
-class CryptoCurrency(
-    val symbol: String,
-    val precision: Int,
-    val feePrecision: Int
-) {
-}
