--- src/main/kotlin/playground/cryptobackedloans/domain/loan/DailyInterest.kt
@@ -1,18 +1,0 @@
-package playground.cryptobackedloans.domain.loan
-
-import playground.cryptobackedloans.domain.common.vo.FiatPrice
-import playground.cryptobackedloans.domain.common.vo.Quantity
-import java.time.LocalDateTime
-
-/**
- * 대출에 발생하는 일별 이자
- */
-class DailyInterest(
-    val id: Long,
-    val quantity: Quantity,
-    val krwPrice: FiatPrice,
-    val usdPrice: FiatPrice,
-    val createdAt: LocalDateTime,
-    val updatedAt: LocalDateTime
-) {
-}
