--- src/main/kotlin/playground/cryptobackedloans/domain/loan/MarginCall.kt
@@ -1,16 +1,0 @@
-package playground.cryptobackedloans.domain.loan
-
-import playground.cryptobackedloans.domain.common.vo.Ltv
-import java.time.LocalDateTime
-
-/**
- * 마진콜
- */
-class MarginCall(
-    val id: Long,
-    val ltv: Ltv, // 마진콜 발동 시점의 ltv
-    val createdAt: LocalDateTime,
-    val releasedAt: LocalDateTime,
-    val updatedAt: LocalDateTime
-) {
-}
