--- src/main/kotlin/playground/cryptobackedloans/domain/user/User.kt
@@ -1,12 +1,0 @@
-package playground.cryptobackedloans.domain.user
-
-import java.time.LocalDateTime
-
-class User(
-    val id: Long,
-    val email: String,
-    val password: String,
-    val createdAt: LocalDateTime,
-    val updatedAt: LocalDateTime
-) {
-}
