--- src/test/kotlin/playground/cryptobackedloans/CryptoBackedLoansApplicationTests.kt
@@ -1,13 +1,0 @@
-package playground.cryptobackedloans
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class CryptoBackedLoansApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-
-}
